Swap (Lng, Lat) to (Lat, Lng) to fix GeoTrellis masking operations.
We copied geojson.js from geotrellis-chatta-demo and noticed some
peculiar behavior. That code is used to serialize polygon layers, created
using the Leaflet draw plugin, in order to POST these shapes to the
backend to be applied as a raster mask.

The problem is that it serializes coordinates in the reverse order from
the standard format that is used in other GeoJSON sources.

We observed that the version of GeoTrellis that we are using expects
GeoJSON coordinates to be swapped which is a deviation from the
standard. This commit resolves that issue by performing a transformation
on each point in a geometry before it is used for masking operations.