Add `crs` parameter to all endpoints to remove ambiguity.
It was confusing that some endpoints expected LatLng projection while
others expected WebMercator. As a result, you wouldn't know which
projection to use without referencing the documentation.

Since it is sometimes convenient to use LatLng, and at other times
more convenient to use WebMercator, I decided to add a parameter to each
endpoint to specify which projection the consumer is using.