WIP - Create tile project
I moved the tile related methods into this project and made it a working
spray service that can be launched via spark-submit

I moved the remaining old code into a legacy folder for the time
being. Useful to have around as a reference while finishing up the new
services, and the tests have not been reconfigured to work with the new setup.