Modernize project and add spark
I have brought in the latest sbt runner to deal with failures resolving
HTTP redirects and updated to a more current version of sbt.

I replaced build.sbt with build.scala, following the lead of
GeoTrellis. I also find it easier to write Scala than to use the sbt
DSL.

I tweaked make-tar to deal with the fact that the sbt assembly plugin now
builds jars into a subdirectory corresponding to the Scala version.

The spark and hadoop dependencies are not used by the existing code, but
will be used by the updated services.