Convert tile service to use Spark
While converting the project to use RasterRDDs and Tiles pulled from S3
I also decomposed the monolithic ModelingServiceSpec into smaller
functional pieces.

I have deferred updating the unit tests in the interest of completing
the conversion sooner, but have left behind all the assertions in a
comment block.

At the time I wrote this conversion the polyMask and polyTileMask
methods could not be converted to use Spark pending a new GeoTrellis
feature. That feature is in the latest master branch, so they can be
implemented in a future commit.