Add scripts to upload jars to Github releases
The `upload-jar-to-release.sh` script is designed to be run on every
Jenkins build of `otm-modeling` after the combined assembly is built. If
the HEAD of the repo matches the most recent tag (sorted by commit date)
the script creates a release for that tag and uploads the compiled jar.

This script uses https://github.com/aktau/github-release to work with
Github releases at a higher level of abstraction.

I broke two smaller helpers out of the main script to ease debugging and testing.