Restore Tile and RasterRDD polygon masking
Masking operations for RDDs were updated in GeoTrellis d63608c so I was
able to easily restore the masking functionality in
LayerMasking.scala. This API is still available in GeoTrellis 0.10.0.

A single tile, specified by zoom level, x, and y, does not have any
geographic information associated with it. GeoTrellis has an API to
identify the cells in tile that are within a specified
polygon (Rasterizer.foreachCellByGeometry), but it requires a geographic
extent as an argument. To obtain the required extent, I have referenced
the code at
http://www.maptiler.org/google-maps-coordinates-tile-bounds-projection/
and added the required helper functions.