Add geocoder and set prioritization bounds based on geocode result
The Google base map already requires the inclusion of the Google Maps API so we
are opting to use it for geocoding as well.

To keep things simple we are always choosing the first result from the geocoder.

We convert a stream of location searches into a stream of lat/lng points and
then convert that point stream into a stream of bounding boxes creating a 10,000
meter wide box centered on the point. This area is roughly city-sized and is the
value that we have been using as the default are for new OpenTreeMap instances.