Add Google place suggestions to location search
Adding `libraries=places` to the script tag that loads the Google Maps API gives
us access to autocomplete functionality that can be easily attached to a
textbox.

In addition to adding the code to handle autocomplete results I moved more of
the code related to geocode result handling into `geocoder.js`.