Add zip code endpoint and connect it to boundary masking UI
When this application was part of OpenTreeMap boundary masks were loaded from
the database. In this standalone version we have opted to load a single boundary
category, Zip Codes, from a flat file in the resources directory.

There are about 30,000 zip codes. Loading and parsing a large GeoJSON file was
triggering out of memory errors even when the heap size was increased to 6GB. We
have opted to keep the memory footprint as small as possible by storing the zip
code and associated GeoJSON geometry in a tab-delimited file that can be read
line-wise. This more efficient method still required more than the default heap
size, so we increased it to 2G in build.sbt.

The zip code file is too large to be committed to the repository so we will
add a build step to copy the file into the resources directory before creating
the service JAR.

The layer masking fails when attempting to mask by a polygon that does not
intersect the breaks bounding box. To avoid this we have added a required `bbox`
parameter to the zip codes endpoint to check for disallowed zip codes and return
a 400 error with a message for the user.

Changing the bounds of the map via the location search box may cause previously
valid zip codes to become invalid. To avoid this, and keep things simple, we
clear the list of zip code masks when we change the bounds of the map.