Show a variant of the welcome dialog when coming from OpenTreeMap
When linking to the demo site from OpenTreeMap, we'd like to continue
to show the welcome dialog (but not the featured cities) and show the
location of the tree map the user came from.

In order to continue to not show the welcome dialog when a user shares a
link, we hide the dialog when the preset parameter is present,
instead of the center parameter as we did previously.

Connects to #181