Add a button to set the prioritization center to the map center
This is based loosely on the way mobile point-of-interest search applications
work. Whenever the user moves the map, we show a button allowing them to set the
center of the prioritization breaks calculation to the current map center.

We moved the centerToBounds function to the prioritization module so it could be
used within that module as well ass the top-level modeling module.